requirements.txt
Flask==3.1.0
httpx==0.28.1
python-dotenv==1.0.1
gunicorn==23.0.0
